requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.1
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
comm==0.2.2
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.1
decorator==5.1.1
efficientnet-pytorch==0.7.1
exceptiongroup==1.2.0
executing==2.0.1
filelock==3.13.1
fonttools==4.49.0
frozenlist==1.4.1
fsspec==2024.2.0
h5py==3.10.0
huggingface-hub==0.21.4
idna==3.6
imageio==2.34.0
ipykernel==6.29.3
ipython==8.22.2
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
lazy_loader==0.3
lightning-utilities==0.10.1
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
munch==4.0.0
nest-asyncio==1.6.0
networkx==3.2.1
nir==1.0.1
nirtorch==1.0
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==24.0
pandas==2.2.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.2.0
pretrainedmodels==0.7.4
prompt-toolkit==3.0.43
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.2.1
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
safetensors==0.4.2
scikit-image==0.22.0
scikit-learn==1.4.1.post1
scipy==1.12.0
segmentation-models-pytorch==0.3.3
six==1.16.0
snntorch==0.8.0
stack-data==0.6.3
sympy==1.12
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.9.2
torch==2.2.1
torchmetrics==1.3.1
torchsummary==1.5.1
torchvision==0.17.1
tornado==6.4
tqdm==4.66.2
traitlets==5.14.2
triton==2.2.0
typing==3.7.4.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
